Fix comment: merge 2 tfm version RemoveRange() back to ImmutableArray_1.cs.
